[optimization]: optimized volume control
